requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-docx==1.1.2
pdfplumber==0.11.4
PyPDF2==3.0.1
spacy==3.7.5
transformers==4.44.2
torch==2.4.1
sentencepiece==0.2.0
openai==1.45.0
langchain==0.3.0
langchain-community==0.3.0
tiktoken==0.7.0
python-dotenv==1.0.1
httpx==0.27.2
pandas==2.2.2
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
httpx==0.27.2
black==24.8.0
ruff==0.6.4
python-magic-bin==0.4.14
streamlit==1.38.0
pandas==2.2.2
plotly==5.24.0
